+++ R/module-define.R
@@ -351,7 +351,11 @@ setMethod(
 ################################################################################
 #' Define a parameter used in a module
 #'
-#' Used to specify a parameter's name, value, and set a default.
+#' Used to specify a parameter's name, value, and set a default. The \code{min} and
+#' \code{max} arguments are ignored by \code{simInit} or \code{spades}; they
+#' are for human use only. To ensure that a user cannot set parameters outside of
+#' a range of values, the module developer should use assertions in their module
+#' code.
 #'
 #' @note Be sure to use the correct NA type: logical (\code{NA}), integer (\code{NA_integer_}),
 #'       real (\code{NA_real_}), complex (\code{NA_complex_}), or character (\code{NA_character_}).
@@ -363,8 +367,14 @@ setMethod(
 #'                  Non-standard evaluation is used for the expression.
 #' @param min       With \code{max}, used to define a suitable range of values.
 #'                  Non-standard evaluation is used for the expression.
+#'                  \emph{These are not tested by} \code{simInit} \emph{or}
+#'                  \code{spades}. These are primarily for human use, i.e., to
+#'                  tell a module user what values the module expects.
 #' @param max       With \code{min}, used to define a suitable range of values.
 #'                  Non-standard evaluation is used for the expression.
+#'                  \emph{These are not tested by} \code{simInit} \emph{or}
+#'                  \code{spades}. These are primarily for human use, i.e., to
+#'                  tell a module user what values the module expects.
 #' @param desc      Text string providing a brief description of the parameter.
 #'
 #' @return data.frame