+++ R/save.R
@@ -12,12 +12,12 @@ doEvent.save <- function(sim, eventTime, eventType, debug = FALSE) {
       #sim <- scheduleEvent(sim, end(sim, sim@simtimes[["timeunit"]]), "save", "end", .last())
     }
     checkPath(sim@paths$outputPath, create = TRUE)
-
   } else if (eventType == "spades") {
     sim <- saveFiles(sim)
   } else if (eventType == "later") {
     sim <- saveFiles(sim)
   } else if (eventType == "end") {
+    sim <- saveFiles(sim)
     message(crayon::green(paste0("Files saved. Use outputs(your simList) for details")), sep = "")
   }
 
@@ -76,8 +76,8 @@ doEvent.save <- function(sim, eventTime, eventType, debug = FALSE) {
 #'
 #' @param sim A \code{simList} simulation object.
 #'
-#' @importFrom dplyr bind_rows
 #' @importFrom data.table data.table
+#' @importFrom dplyr bind_rows
 #' @export
 #' @rdname saveFiles
 #'